Added WebSocket code, resolved some issues while updating board....Real-time changes still in progress

+++ workboard-backend/src/main/java/com/project/workboard/service/BoardService.java
@@ -11,11 +11,13 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.project.workboard.dto.ApiResponseDTO;
 import com.project.workboard.dto.BoardDataDTO;
+import com.project.workboard.dto.BoardEvent;
 import com.project.workboard.dto.BoardWithMembersProjection;
 import com.project.workboard.dto.SavedBoardDataDTO;
 import com.project.workboard.dto.SavedBoardDataDTO.MemberDataDTO;
@@ -23,6 +25,7 @@ import com.project.workboard.dto.SavedBoardMemberDTO;
 import com.project.workboard.dto.SavedBoardMemberProjection;
 import com.project.workboard.dto.SavedTaskMemberDTO;
 import com.project.workboard.dto.TaskDataDTO;
+import com.project.workboard.dto.UpdatedBoardSummaryDTO;
 import com.project.workboard.entity.AppUser;
 import com.project.workboard.entity.Board;
 import com.project.workboard.entity.BoardList;
@@ -50,6 +53,13 @@ public class BoardService {
 
 	@Autowired
 	private JwtService jwtService;
+
+	@Autowired
+	private SimpMessagingTemplate messagingTemplate;
+
+	public BoardService(SimpMessagingTemplate messagingTemplate) {
+		this.messagingTemplate = messagingTemplate;
+	}
 
 	public ResponseEntity<?> saveBoard(BoardDataDTO boardData) {
 		System.out.println("inside BoardService's createBoard, boardData: " + boardData.toString());
@@ -167,39 +177,51 @@ public class BoardService {
 		System.out.println("Inside BoardService :: updateBoard, boardData: " + boardData.toString());
 
 		try {
-
 			// Getting board
 			Optional<Board> boardOpt = boardRepository.findById(boardData.getBoardId());
 			if (boardOpt.isEmpty()) {
 				throw new IllegalArgumentException("Board not found");
 			}
 			Board board = boardOpt.get();
+			System.out.println("board: " + board.toString());
 
 			// checking name boolean
 			boolean boardChanged = false;
 			if (!Objects.equals(board.getName(), boardData.getName())) {
+				System.out.println("board-name updated to : " + boardData.getName());
 				board.setName(boardData.getName());
 				boardChanged = true;
 			}
 
 			// checking description
-			boardChanged = false;
 			if (!Objects.equals(board.getDescription(), boardData.getDescription())) {
+				System.out.println("board-description updated to : " + boardData.getDescription());
 				board.setDescription(boardData.getDescription());
 				boardChanged = true;
 			}
 
 			// checking members
+			boolean memberChanged = false;
 			BoardDataDTO.Member[] members = boardData.getMembers();
 			if (members.length > 0) {
 				// Checking if updates are made to Board-members
-				saveUpdatesForBoardMembers(boardData, board);
+				memberChanged = saveUpdatesForBoardMembers(boardData, board);
+				if (memberChanged)
+					boardChanged = true;
 			}
 
 			String msg = "Board data & members saved successfully";
 			if (boardChanged) {
-				// saving the updates to the board
+				// SAVING the updates to the board
 				board = boardRepository.save(board);
+
+				// NOTIFYING OTHER USERS ABOUT THE UPDATE
+				UpdatedBoardSummaryDTO boardSummary = new UpdatedBoardSummaryDTO(board.getId(), board.getName(),
+						board.getDescription());
+
+				// BROADCAST to all viewers of this board
+				notifyBoardViewers(board.getId(), boardSummary, 1);
+
 				msg = "Successfully saved updates to board with id: " + board.getId() + " & name: " + board.getName();
 				System.out.println(msg);
 			} else {
@@ -223,7 +245,43 @@ public class BoardService {
 		}
 	}
 
-	private void saveUpdatesForBoardMembers(BoardDataDTO boardData, Board board) {
+	private void notifyBoardViewers(int boardId, Object payload, int flag) {
+		BoardEvent evt = new BoardEvent();
+		evt.setType((flag == 1) ? BoardEvent.Type.UPSERT : BoardEvent.Type.DELETE);
+		evt.setBoardId(boardId);
+		evt.setPayload(payload);
+		evt.setVersion(System.currentTimeMillis());
+
+		String dest = "/topic/board." + boardId;
+		messagingTemplate.convertAndSend(dest, evt);
+	}
+
+	public ResponseEntity<?> testEvent() {
+		System.out.println("Inside BoardService :: testEvent");
+		try {
+			BoardEvent evt = new BoardEvent();
+			evt.setType(BoardEvent.Type.UPSERT);
+			evt.setBoardId(10);
+
+			Map<String, Object> payload = new HashMap<>();
+			payload.put("id", 10);
+			payload.put("name", "Test Name");
+			payload.put("description", "Updated via test");
+			evt.setPayload(payload);
+
+			messagingTemplate.convertAndSend("/topic/board.10", evt);
+
+			// Data is saved successfully, let's send Api-Response for the same
+			ApiResponseDTO<Integer> apiResponse = new ApiResponseDTO<>(true, 1, "testing");
+			return ResponseEntity.ok(apiResponse);
+		} catch (Exception e) {
+			System.out.println("Exception while testing live-updates via socket-connection: " + e.getMessage());
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+					.body("Error while testing live-updates via socket-connection");
+		}
+	}
+
+	private boolean saveUpdatesForBoardMembers(BoardDataDTO boardData, Board board) {
 		System.out.println("inside BoardService :: saveUpdatesForBoardMembers");
 
 		// getting existing members of the task
@@ -273,8 +331,8 @@ public class BoardService {
 
 				// add to removeList
 				removeList.add(boardMember);
-
-				memberChanged = true;
+				
+//				memberChanged = true;
 			} else {
 				// existingMember is present in updatedMemberMap
 				// let's check for change in role
@@ -289,7 +347,7 @@ public class BoardService {
 					// add to updateList
 					updateList.add(boardMember);
 
-					memberChanged = true;
+//					memberChanged = true;
 				}
 			}
 		}
@@ -319,33 +377,43 @@ public class BoardService {
 				// pushing to addList
 				addList.add(boardMember);
 
-				memberChanged = true;
+//				memberChanged = true;
 			}
 		}
 
 		// saving the updates to Db
-		System.out.println("board-members to add: ");
-		for (BoardMember bm : addList) {
-			System.out.println(bm.toString());
+		if (addList.size() > 0) {
+			System.out.println("board-members to add: ");
+			for (BoardMember bm : addList) {
+				System.out.println(bm.toString());
+			}
+			boardMemberRepository.saveAll(addList);
+			memberChanged = true;
 		}
-		boardMemberRepository.saveAll(addList);
 
-		System.out.println("board-members to update: ");
-		for (BoardMember bm : updateList) {
-			System.out.println(bm.toString());
+		if (updateList.size() > 0) {
+			System.out.println("board-members to update: ");
+			for (BoardMember bm : updateList) {
+				System.out.println(bm.toString());
+			}
+			boardMemberRepository.saveAll(updateList);
+			memberChanged = true;
 		}
-		boardMemberRepository.saveAll(updateList);
 
-		System.out.println("board-members to remove: ");
-		for (BoardMember bm : removeList) {
-			System.out.println(bm.toString());
+		if (removeList.size() > 0) {
+			System.out.println("board-members to remove: ");
+			for (BoardMember bm : removeList) {
+				System.out.println(bm.toString());
+			}
+			boardMemberRepository.deleteAll(removeList);
+			memberChanged = true;
 		}
-		boardMemberRepository.deleteAll(removeList);
 
 		if (memberChanged)
 			System.out.println("Updates to board-members are saved successfully");
 		else
 			System.out.println("No changes were made to the board-members of the board");
+		return memberChanged;
 	}
 
 	@Transactional
